backend/requirements.txt
fastapi 
uvicorn
python-dotenv
pydantic
psycopg2-binary
cloud-sql-python-connector[pg8000]==1.12.0
sqlalchemy
sqlmodel
python-multipart 
fastapi==0.60.0
uvicorn==0.11.0
requests==2.19.0
PyYAML==5.1
Django==2.2.0      

# Google Cloud Services
google-cloud-storage
google-cloud-vision
google-generativeai

# Google ADK for Agent Development
google-adk

# Authentication
google-auth
PyJWT
httpx

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0
hypothesis>=6.82.0
